[Refac] PhotonStageSceneRoomManager 클래스 리팩토링
