+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68145c9e2136344a481c478170a14fb8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
