+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState.cs
@@ -1,0 +1,13 @@
+using Photon.Pun;
+
+public abstract class StageState : MonoBehaviourPun
+{
+    protected void Awake()
+    {
+        AddToDictionary();
+    }
+
+    public abstract void AddToDictionary();
+    
+    public abstract void Action();
+}
