+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b93572cbc6e44b2a9e020dfa82ba37c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
