+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_GameCompletion.cs
@@ -1,0 +1,54 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Cysharp.Threading.Tasks;
+using Photon.Pun;
+using Util.Helper;
+
+public class StageState_GameCompletion : StageState
+{
+    public override void AddToDictionary()
+    {
+        StageManager.Instance.StageDataManager.SequenceActionDictionary[StageDataManager.StageSequence.GameCompletion] = this;
+    }
+
+    public override void Action()
+    {
+        if (!PhotonNetwork.IsMasterClient)
+            return;
+
+        AsyncEndProductionScheduler().Forget();
+    }
+
+    private async UniTaskVoid AsyncEndProductionScheduler()
+    {
+        await UniTask.Delay(TimeSpan.FromSeconds(2f), DelayType.UnscaledDeltaTime);
+
+        ScoreCalculateHelper.Calculate();
+        UpdatePlayerRanking();
+
+        await UniTask.Delay(TimeSpan.FromSeconds(1f), DelayType.UnscaledDeltaTime);
+
+        photonView.RPC("RpcSetRoundState", RpcTarget.All);
+    }
+
+    private void UpdatePlayerRanking()
+    {
+        List<KeyValuePair<int, PlayerData>> sortedPlayers =
+            StageManager.Instance.PlayerContainer.PlayerDataByIndex.OrderByDescending(pair => pair.Value.Score)
+                .ToList();
+
+        StageManager.Instance.PlayerContainer.CachedPlayerIndicesForResults.Clear();
+
+        foreach (KeyValuePair<int, PlayerData> pair in sortedPlayers)
+        {
+            StageManager.Instance.PlayerContainer.CachedPlayerIndicesForResults.Add(pair.Key);
+        }
+    }
+
+    [PunRPC]
+    public void RpcSetRoundState()
+    {
+        StageManager.Instance.StageDataManager.SetSequence(StageDataManager.StageSequence.RoundCompletion);
+    }
+}
