+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_GameCompletion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d690b74891c284eb8a6f71369823dfb4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
