+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_GameInProgress.cs
@@ -1,0 +1,12 @@
+public class StageState_GameInProgress : StageState
+{
+    public override void AddToDictionary()
+    {
+        StageManager.Instance.StageDataManager.SequenceActionDictionary[StageDataManager.StageSequence.GameInProgress] = this;
+    }
+
+    public override void Action()
+    {
+        
+    }
+}
