+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_GameInProgress.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ecb62271d679e445b92ea4bf2729cd37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
