+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_PlayersReady.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6100642679f246e68795e94ffe739aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
