+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_RoundCompletion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36362a9c84f4c4efbae92ea991d3bc3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
