+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_Transitioning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01c20c8af391442688315b947899ead6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
