+++ JKC_Fallguys/Assets/1_Scripts/Stage/StageState/StageState_WaitingForPlayers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6487d9593bae849f6a189d67f862f3ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
