+++ JKC_Fallguys/Assets/1_Scripts/Util/Helper/ScoreCalculateHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa7ea693515d04a47adb2e082d1e4c97
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
